Add present queue
Also cleaned up and organized some debug information.

I was supposed to check if the drawing and presenting devices were different and make separate info for them. I could not figure out how to get that working in Odin with the information I had in the tutorial. So for now I will just handle a single device (most common case anyway).